# api/openapi.yaml
openapi: 3.0.3
info:
  title: Vesta Kubernetes API
  description: |
    REST API for the Vesta Kubernetes PaaS platform.
    Manages pipelines, apps, deployments, secrets, and templates.
  version: 0.1.0
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: /api/v1
    description: Vesta API v1
security:
  - BearerAuth: []
...
paths:
...
  /pipelines/{pipelineId}/apps:
    parameters:
      - $ref: "#/components/parameters/PipelineId"
    post:
      tags: [Apps]
      summary: Create an app in a pipeline stage
      operationId: createApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppCreate"
      responses:
        "201":
          description: App created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
  /apps:
    get:
      tags: [Apps]
      summary: List apps
      operationId: listApps
      parameters:
        - name: pipeline
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: team
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: List of apps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppList"
  /apps/{appId}:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Apps]
      summary: Get app details and status
      operationId: getApp
      responses:
        "200":
          description: App details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Apps]
      summary: Update app configuration
      operationId: updateApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUpdate"
      responses:
        "200":
          description: App updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
    delete:
      tags: [Apps]
      summary: Delete app
      operationId: deleteApp
      responses:
        "204":
          description: App deleted

  # ── Deployments ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PipelineId:
      name: pipelineId
      in: path
      required: true
      schema:
        type: string
    AppId:
      name: appId
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string

    # ── Pipeline schemas ─────────────────────────────────────────────────
...
    AppCreate:
      type: object
      required: [name, stage]
      properties:
        name:
          type: string
        stage:
          type: string
        git:
          type: object
          properties:
            provider:
              type: string
            repository:
              type: string
            branch:
              type: string
            autoDeployOnPush:
              type: boolean
        build:
          type: object
          properties:
            strategy:
              type: string
              enum: [runpacks, buildpacks, dockerfile, nixpacks, image]
            dockerfile:
              type: string
        image:
          $ref: "#/components/schemas/ImageConfig"
        runtime:
          $ref: "#/components/schemas/RuntimeConfig"
        scaling:
          $ref: "#/components/schemas/ScalingConfig"
        resources:
          $ref: "#/components/schemas/ResourceConfig"
        ingress:
          $ref: "#/components/schemas/IngressConfig"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/AddonConfig"
        customConfig:
          type: object
          description: Custom Kubernetes configuration (labels, annotations, podSpec, containerSpec, extraResources)
    AppUpdate:
      type: object
      properties:
        image:
          $ref: "#/components/schemas/ImageConfig"
        runtime:
          $ref: "#/components/schemas/RuntimeConfig"
        scaling:
          $ref: "#/components/schemas/ScalingConfig"
        resources:
          $ref: "#/components/schemas/ResourceConfig"
        ingress:
          $ref: "#/components/schemas/IngressConfig"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/AddonConfig"
        customConfig:
          type: object
    App:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        pipeline:
          type: string
        stage:
          type: string
        image:
          $ref: "#/components/schemas/ImageConfig"
        runtime:
          $ref: "#/components/schemas/RuntimeConfig"
        scaling:
          $ref: "#/components/schemas/ScalingConfig"
        resources:
          $ref: "#/components/schemas/ResourceConfig"
        ingress:
          $ref: "#/components/schemas/IngressConfig"
        status:
          $ref: "#/components/schemas/AppStatus"
        createdAt:
          type: string
          format: date-time
    AppList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/App"
        total:
          type: integer
    AppStatus:
      type: object
      properties:
        phase:
          type: string
          enum: [Pending, Building, Deploying, Running, Failed, Sleeping]
        url:
          type: string
        currentImage:
          type: string
        lastDeployedAt:
          type: string
          format: date-time
        currentReplicas:
          type: integer
        desiredReplicas:
          type: integer
    ImageConfig:
      type: object
      required: [repository]
      properties:
        repository:
          type: string
        tag:
          type: string
        pullPolicy:
          type: string
          enum: [Always, IfNotPresent, Never]
        imagePullSecrets:
          type: array
          items:
            $ref: "#/components/schemas/NameRef"
    RuntimeConfig:
      type: object
      properties:
        port:
          type: integer
        command:
          type: string
        args:
          type: array
          items:
            type: string
        env:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
        secrets:
          type: array
          items:
            type: object
            description: Secret binding (secretRef, secretKeyRef, or secretMount)
    ScalingConfig:
      type: object
      properties:
        replicas:
          type: integer
        autoscale:
          type: object
          properties:
            enabled:
              type: boolean
            minReplicas:
              type: integer
            maxReplicas:
              type: integer
            metrics:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [cpu, memory, custom]
                  name:
                    type: string
                  targetAverageUtilization:
                    type: integer
                  targetAverageValue:
                    type: string
    ResourceConfig:
      type: object
      properties:
        size:
          type: string
        requests:
          type: object
          properties:
            cpu:
              type: string
            memory:
              type: string
        limits:
          type: object
          properties:
            cpu:
              type: string
            memory:
              type: string
    IngressConfig:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
        tls:
          type: boolean
        clusterIssuer:
          type: string
        basicAuth:
          type: boolean
        annotations:
          type: object
          additionalProperties:
            type: string
    AddonConfig:
      type: object
      required: [type]
      properties:
        type:
          type: string
        version:
          type: string
        size:
          type: string

    # ── Deploy schemas ───────────────────────────────────────────────────
...
    NameRef:
      type: object
      required: [name]
      properties:
        name:
          type: string
...